deprecate StringUtil and do a minor refactor
