+++ src/main/java/tv/isshoni/winry/api/WinryContext.java
@@ -1,8 +1,9 @@
-package tv.isshoni.winry.api.entity.context;
+package tv.isshoni.winry.api;
 
 import tv.isshoni.araragi.logging.AraragiLogger;
 import tv.isshoni.araragi.stream.Streams;
 import tv.isshoni.winry.api.annotation.Bootstrap;
+import tv.isshoni.winry.api.entity.context.IWinryContext;
 import tv.isshoni.winry.api.entity.executable.IExecutable;
 import tv.isshoni.winry.entity.annotation.IWinryAnnotationManager;
 import tv.isshoni.winry.entity.annotation.inject.IInjectionRegistry;
