+++ src/main/java/tv/isshoni/winry/api/bootstrap/WinryEventExecutable.java
@@ -1,5 +1,6 @@
-package tv.isshoni.winry.api.entity.event;
+package tv.isshoni.winry.api.bootstrap;
 
+import tv.isshoni.winry.api.entity.event.IEvent;
 import tv.isshoni.winry.api.entity.executable.IExecutable;
 import tv.isshoni.winry.entity.event.IEventBus;
