+++ src/main/java/tv/isshoni/winry/api/entity/context/IContextual.java
@@ -1,5 +1,7 @@
 package tv.isshoni.winry.api.entity.context;
 
+import tv.isshoni.winry.api.WinryContext;
+
 import java.util.Optional;
 
 public interface IContextual {
