+++ src/main/java/tv/isshoni/winry/internal/annotation/processor/parameter/WinryContextProcessor.java
@@ -3,7 +3,7 @@ package tv.isshoni.winry.internal.annotation.processor.parameter;
 import tv.isshoni.araragi.annotation.model.IParameterSupplier;
 import tv.isshoni.winry.api.annotation.parameter.Context;
 import tv.isshoni.winry.api.entity.context.IWinryContext;
-import tv.isshoni.winry.api.entity.context.WinryContext;
+import tv.isshoni.winry.api.WinryContext;
 
 import java.lang.reflect.Parameter;
 import java.util.Objects;
